requirements.txt
Flask>=3.0.0
gspread>=5.12.0
oauth2client>=4.1.3
gunicorn>=21.2.0
Flask-APScheduler>=1.13.0
line-bot-sdk>=3.11.0

